pyproject.toml
[project]
name = "codara"
version = "0.1.0"
description = "Codara: AI Agent Cli for Coding"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
]

[project.scripts]
codara = "app.cli:main"

[tool.setuptools]
packages = ["app"]